% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtils.r
\name{downloadFiles}
\alias{downloadFiles}
\title{downloadFiles}
\usage{
downloadFiles(datalake, folder, token)
}
\arguments{
\item{datalake}{String the data lake name.}

\item{folder}{String the subfolder name.}

\item{token}{String the Bearer token.}
}
\value{
data.table
}
\description{
Downloads all files in a datalake folder. Assumes files are ndjson format.
}
