% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtils.r
\name{getFile}
\alias{getFile}
\title{getFile}
\usage{
getFile(datalake, folder, filename, token)
}
\arguments{
\item{datalake}{String the data lake name.}

\item{folder}{String the subfolder name.}

\item{filename}{String the file name.}

\item{token}{String the Bearer token.}
}
\value{
String text contents of a file.
}
\description{
getFile
}
