% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtils.r
\name{getToken}
\alias{getToken}
\title{getToken}
\usage{
getToken(tenantID, clientID, apiKey)
}
\arguments{
\item{tenantID}{String the tenant Id.}

\item{clientID}{String the client Id.}

\item{apiKet}{String the API secret.}
}
\value{
String the access token
}
\description{
getToken
}
