% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtils.r
\name{listFiles}
\alias{listFiles}
\title{listFiles}
\usage{
listFiles(datalake, folder, token)
}
\arguments{
\item{datalake}{String the data lake name.}

\item{folder}{String the subfolder name.}

\item{token}{String the Bearer token.}
}
\value{
List pathSuffix for each file in the folder.
}
\description{
listFiles
}
