% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtils.r
\name{listFolder}
\alias{listFolder}
\title{listFolder}
\usage{
listFolder(datalake, folder, token)
}
\arguments{
\item{datalake}{String the data lake name.}

\item{folder}{String the subfolder name.}

\item{token}{String the Bearer token.}
}
\value{
List complete contents of a folder.
}
\description{
listFolder
}
